Hackney doesn't accept nil as body
This wasn't causing issues before because we were using other
representations for an empty body ("", [] or %{}) which meant this
branch of the pattern matching was never executed.

With the introduction of GH-42
(https://github.com/dnsimple/dnsimple-elixir/pull/42) the internal
representation of an empty body was set to nil, which in turn caused a
bug because Hackney doesn't accept nil as a value for the body.