Example of decoders with Poison
WARN: this commit is just here to trigger a discussion and WILL BE
REVERTED.

After this comment
https://github.com/dnsimple/dnsimple-elixir/pull/49#issuecomment-245622102
I decided to give Poison decoders a try. I did not take them into
consideration on the first place because I thought it was going to be
tricky to change all the parsing code to use them. However, as you can
see in the test I just added, is pretty straightforward.